requirements.txt
fastapi
uvicorn
crewai==0.80.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
langchain>=0.3.0
requests
python-multipart
jinja2